requirements.txt
alembic==1.6.2
bcrypt==3.2.0
boto3==1.20.34
botocore==1.23.34
cffi==1.15.0
click==8.0.0
cryptography==36.0.1
Flask==2.0.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-S3==0.3.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jwt==1.3.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.3.23
urllib3==1.26.8
Werkzeug==2.0.1
zipp==3.7.0
